+++ UserLib/board_main.hpp
@@ -11,12 +11,11 @@
 #include "board_params.hpp"
 #include "motor.hpp"
 
+#include "CommonLib/MotorMath/sin_table.hpp"
 #include "CommonLib/pwm.hpp"
 #include "CommonLib/pid.hpp"
-#include "CommonLib/math.hpp"
 #include "CommonLib/programable_LED.hpp"
 #include "CommonLib/LED_pattern.hpp"
-#include "CommonLib/sincos_enc.hpp"
 #include "CommonLib/cordic.hpp"
 #include "CommonLib/fdcan_control.hpp"
 
@@ -42,9 +41,7 @@ namespace LMDBoard{
 	inline auto motor = LMDLib::Motor{
 		SabaneLib::PWMHard{&htim1,TIM_CHANNEL_2},
 		SabaneLib::PWMHard{&htim1,TIM_CHANNEL_3},
-		SabaneLib::PWMHard{&htim1,TIM_CHANNEL_1},
-		[](q15_t r)->SabaneLib::MotorMath::SinCos {return table.sin_cos(r);},
-		atan_enc
+		SabaneLib::PWMHard{&htim1,TIM_CHANNEL_1}
 	};
 
 	namespace PIDIns{
